+++ RealmSwift/PersistedProperty.swift
@@ -297,7 +297,7 @@ extension Persisted: OptionalCodingWrapper where Value: ExpressibleByNilLiteral 
  are valid), optional enum properties will return `nil`, and non-optional
  properties will abort the process.
  */
-public protocol PersistableEnum: _PersistableInsideOptional, RawRepresentable, CaseIterable, RealmEnum, RealmCollectionValue, MinMaxType, Comparable where RawValue: Comparable {
+public protocol PersistableEnum: _PersistableInsideOptional, RawRepresentable, CaseIterable, RealmEnum, _RealmCollectionValueInsideOptional, MinMaxType, Comparable where RawValue: Comparable {
 }
 
 extension PersistableEnum {
@@ -307,13 +307,25 @@ extension PersistableEnum {
     public static func < (lhs: Self, rhs: Self) -> Bool {
         return lhs.rawValue < rhs.rawValue
     }
+    /// :nodoc:
+    public static func _rlmDefaultValue() -> Self {
+        Self.allCases.first!
+    }
 }
 
+#if swift(>=5.5)
+/// A type which can be indexed.
+///
+/// This protocol is merely a tag and declaring additional types as conforming
+/// to it will simply result in runtime errors rather than compile-time errors.
+@_marker public protocol _Indexable {}
+#else
 /// A type which can be indexed.
 ///
 /// This protocol is merely a tag and declaring additional types as conforming
 /// to it will simply result in runtime errors rather than compile-time errors.
 public protocol _Indexable {}
+#endif
 
 extension Persisted where Value.PersistedType: _Indexable {
     /// Declares an indexed property which is lazily initialized to the type's default value.
@@ -326,11 +338,19 @@ extension Persisted where Value.PersistedType: _Indexable {
     }
 }
 
+#if swift(>=5.5)
+/// A type which can be made the primary key of an object.
+///
+/// This protocol is merely a tag and declaring additional types as conforming
+/// to it will simply result in runtime errors rather than compile-time errors.
+@_marker public protocol _PrimaryKey {}
+#else
 /// A type which can be made the primary key of an object.
 ///
 /// This protocol is merely a tag and declaring additional types as conforming
 /// to it will simply result in runtime errors rather than compile-time errors.
 public protocol _PrimaryKey {}
+#endif
 
 extension Persisted where Value.PersistedType: _PrimaryKey {
     /// Declares the primary key property which is lazily initialized to the type's default value.